+++ web/portal.php
@@ -10,8 +10,7 @@ use fkooman\Config\Config;
 use fkooman\VpnPortal\PdoStorage;
 use fkooman\VpnPortal\VpnPortalService;
 use fkooman\VpnPortal\VpnCertServiceClient;
-use fkooman\Rest\Plugin\BasicAuthentication;
-use fkooman\Rest\Plugin\MellonAuthentication;
+use fkooman\Rest\Plugin\Mellon\MellonAuthentication;
 use GuzzleHttp\Client;
 
 set_error_handler(
@@ -30,24 +29,19 @@ try {
         $config->s('PdoStorage')->l('username', false),
         $config->s('PdoStorage')->l('password', false)
     );
+
+    // Database
     $pdoStorage = new PdoStorage($pdo);
 
-    if ('BasicAuthentication' === $config->l('authType', true)) {
-        $authPlugin = new BasicAuthentication(
-            $config->s('basicAuthentication', true)->l('userId', true),
-            $config->s('basicAuthentication', true)->l('userPass', true)
-        );
-    } elseif ('MellonAuthentication' === $config->l('authType', true)) {
-        $authPlugin = new MellonAuthentication(
-            $config->s('mellonAuthentication')->l('samlAttribute', true)
-        );
-    } else {
-        throw new InternalServerErrorException("unsupported authentication type");
-    }
+    // Authentication
+    $mellonAuthentication = new MellonAuthentication(
+        $config->l('mellonAttribute', true)
+    );
 
-    $serviceUri = $config->s('vpnCertService', true)->l('serviceUri', true);
-    $serviceAuth = $config->s('vpnCertService', true)->l('serviceUser', true);
-    $servicePass = $config->s('vpnCertService', true)->l('servicePass', true);
+    // VPN Certificate Service Configuration
+    $serviceUri = $config->s('VpnCertService', true)->l('serviceUri', true);
+    $serviceAuth = $config->s('VpnCertService', true)->l('serviceUser', true);
+    $servicePass = $config->s('VpnCertService', true)->l('servicePass', true);
 
     $client = new Client(
         array(
@@ -58,8 +52,14 @@ try {
     );
 
     $vpnCertServiceClient = new VpnCertServiceClient($client, $serviceUri);
-    $request = Request::fromIncomingRequest(new IncomingRequest());
-    $vpnPortalService = new VpnPortalService($pdoStorage, $authPlugin, $vpnCertServiceClient);
+
+    $vpnPortalService = new VpnPortalService($pdoStorage, $vpnCertServiceClient);
+    $vpnPortalService->registerBeforeEachMatchPlugin($mellonAuthentication);
+
+    $request = Request::fromIncomingRequest(
+        new IncomingRequest()
+    );
+
     $vpnPortalService->run($request)->sendResponse();
 } catch (Exception $e) {
     if ($e instanceof HttpException) {