+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.3
+- update the default user inteface a little
+- update to new `fkooman/rest`
+- use `fkooman/rest-plugin-mellon` now for authentication
+- update code for new dependencies
+
 ## 0.1.2
 - implement CSRF fix
 - add CSP and X-Frame-Options headers to default Apache config